I bet you weren't expecting this update!
I know it has been a while.
As I expressed in the previous update, back in October, and a week straight of working on this project, I did get a bit fed up of it. But I don't like leaving things so close to completion, but not finished. So Two weeks ago, I came back to it and did a bit of work on the weekends. I have now fixed the soft-lock thing where if someone disconnects, the entire game now stops, rather than just waiting for that person to go. I have also added a back button. There were also some issues with the scaling of the window when you weren't in 1920x1080 and 100% scaling. 
If you have decided to look at the source code, you may have spotted what I am thinking of trying. If not, look at line 507. I may do it, but probably won't as the idea of planning that and making it look good no matter what the dimensions of the window are is really daunting for me.
We'll see. Who knows, if you wait another year, maybe I'll update this again with that feature!

Best, 
id-hkj